requirements.txt
streamlit==1.39.0
torch
torchvision
pillow
